Starts program refactoring (#1)
This PR starts refactoring the program. It creates classes to encapsulate logics
and manage calculations in a more elegant way. It also deprecates the
passing-through of unnamed dicts for classes and typed dicts with clear
names that reveal the intention and usage of things.

Additionally, the repo setup is cleaned up. I am removing tools that are
not in use (e.g. SQLFluff) and updating some local VS Code configurations.