Cities data (#2)
This PR introduces the cities data. For now, only the Unification of
Italy campaign is included. The PR introduces the unit tests neccessary
to validate the YAML file with all the data.

The PR modifies the user's access point. It no longer requires users to
pass information on city's rss potential. It also adds the calculations
for the city effects and the displaying of the effects.