Re-organize modules (#12)
This PR re-organizes the modules. It introduces new modules for
`geo_features`, `resources`, and `effects`. Each of these modules
handles all classes and logics for the respective data-entity. For each
of them I am adding:

- an Enum class for the different types or rss, geos, and effects
- a typed dict for helping with typing data reads
- a dataclass for intializing with defaults in actual program logics

The `city_data` module is merged into the `city` module as there's no
longer any need to have both. The `buildings` module is renamed
`building`.