Add `Scenario` class (#25)
This PR adds the `Scenario` class. This class takes two possible
building configurations for a given city in a given campaign and
displays the results side by side for easy visualization. It also
changes the `City.build_results_display()` method from private to public
because the `Scenario` class needs to access it.