Fix table alignment (#30)
This PR fixes the misalignment between tables that compare two different
scenarios and have different number of buildings.

The solution is not the cleanest though. Because of how `rich` works, we
end up having the height calculations in two places, which introduces
code repetition. Even worst, in the case of the `Scenario` class we need
to do the calculations in the `display_scenario_results()` method,
wheras in the `City` class we need to do the calculations in the
`build_city_display()` method. This raises the cognitive load by having
the "same" action being the resposability of different methods.

This highlights the need for introducing a `DisplayResults` class. See
#31 for that.