Add cities from the Conquest of Britain campaign (#33)
Add cities for the Conquest of Britain campaign. Unfortunately, in this
campaign, the city "Durobrivae" appears twice. The one in the central
marshes of Britain is here named "Durobrivae II".

Before adding the cities for another campaigm, explore the posibility of
having CSpell dictionaries for this things. The amount of names is
cluttering the CSpell config. See #34 for this.