Expand helpers functionality (#53)
This PR expands the behavior of the helper dataclasses
`ResourceCollection`, `GeoFeatures`, and `EffectBonuses` so that they
have dict-like methods.

Adds the `__iter__` dunder method for iterating over "keys" (the members
of the class).

Adds the `items()` method for iterating over "key-value pairs". Adds the
`values()` method for iterating over the values. Notice that this
methods differ from the dict implementation.

Adds the `get()` method for getting the value of a member using the
member's name. For example, `.get("food")` returns the amount of food.
This method diviates from the dict method. in the following ways:

- Invalid key -> raise error
- No default value accepted

The reason for these changes is that this clases are not general classes
that could potentially take any member (key). They are specifically
disigned for the needs to this program. Furthermore, they are always
init'ed with a default value (zero).

---
These methods are going to be very useful for #28.