Improve `City` class declaration (#68)
This PR improves the blueprint of the `City` class by passing additional
args to the different `field`s declarations. With this, the repr and
sorting of cities is propertly coded. In a nutshell, both repr and sort
should be handled by the campaign and name fields. Other ways of sorting
might be needed in other parts of the program (for example, the
`Kingdom` class sorts based on focus and name), but that is of no
concern to the `City` class itself.