Updates defenses (#72)
This PR updates how the defenses sub-fields get computed.

Although the current code is how I would like all of these operations to
look like, this update is done because a) this only works because the
sub-fields do not depend on each other, and b) code symetry, I consider
more imporant that all sub-field calculations in the `__post_init__()`
look the same.