Refactors the `Building` class (#79)
This PR refactors the `Building` class. Buildings are no longer being
created all at once and placed in a dictionary. Instead, each instance
of the `City` class will have its own buildings. In short, cities are
now collections of buildings with some additional properties and
mechanics.

The new `Building` class comes already equiped with `add_workers()`,
`remove_workers()`, and `set_workers()` methods that will be needed for
#16.

A new `building_example.py` is created with examples of buildings and
the workers' methods.

This refactoring triggers a refactoring of the `City` class too. This
class will now take a `list[Building]` buy default. A new classmethod is
added to create cities from a `BuildingsCount`. Additioanlly, the class'
public API is completely rebuilt as it requires different ways to
working with its internal state now that said state is based instances
on other objects and not just a simple dict.

New methods have been added to the `City` class for `get_building()`,
`has_building()`, `get_buildings_count()`. The method
`_validate_unknown_buildings()` is removed as this validation is now
part of the `Building` class (any unknown building just cannot be
created and raises an error.

Additioanl tests are included for both the `Buidling` and `City` classes
to test the new instantiation of these objects.

---
Other modules like display, kingdom, or scenario received just small
twicks. Changes here are limited to the adjustments needed to work with
the new `City` class' API.