Moves the `CityDisplay` class to the `city` module (#88)
This PR takes care of moving the `CityDisplay` class to the `city`
module. Since this works perfectly fine, this will be the whole scope of
this PR. Additional PRs will evaluate whether it is convinient to merge
them or not.

This PR also adds unit tests for the city display methods that do the
calculations.