Adds display for buildings (#91)
This PR adds the `Building.display_building()` public method. This
method displays the building's information to the terminal. This method
replaces the `Building.show()` method, which is here removed.

<img width="953" height="497" alt="image"
src="https://github.com/user-attachments/assets/02b24202-0d62-4484-8b14-d69274c1bc4b"
/>

Information on the required building(s) is not included here because
there's a bug with the data. See #90 for more on this.