Refactors the `Scenario` class (#93)
This PR refactors the `Scenario` class. The main objective here is to
decouple the `Scenario` and the `CityDisplay` classes.

This is a better design not only because less coupling = better
maintainability. But also because this particular coupling doesn't make
sense. After #90, the instances of the `City` class already know about
the `CityDisplay`. Therefore, for the `Scenario` class to have both
instances of `City` and `CityDisplay` classes just doesn't make sense.
It is a duplication.

This PR starts by splitting the `City.display_city()` method. Currently,
the method is building the displayer and displaying the results. Now,
`City.build_city_displayer()` will build and return a displayer. Then
`City.display_city()` will only be responsible of displaying it.

With this change, the `Scenario` class no longer needs to handle
`DisplayCity` objects. Instead, it now calls the display builder for
each `City` instance that it contains, but passes the configuration that
itself built for all cities in the scenario. Thanks to these changes,
the members `Scenario.cities_display: list[CityDisplay]` and
`Scenario._build_cities_display()` are deprecated and removed.

With all of this done, the `CityDisplay` class is now internal to the
`city` module. Therefore, it is here renamed to `_CityDisplay`.

Documentation is updated accordingly.