Modeling of the Supply Dump building (#99)
This PR solves the problem of modeling the Supply Dump established in
#36. This is just one idea on how to do it. It might be changed later.

This PR introduces the flag to the cities data. True for those that have
it, false otherwise. It also adds the modeling to the `City` class. A
boolean member is added to mark those cities that have a Supply Dump.
The necessary validations are added too. The user does not need to
supply the Supply Dump as a building. The program will add it the user
does not.

This PR also adds unit tests for city scenarios with Supply Dump.