Fix required building (#100)
This PR fixes the issues with the required building.

### Changes to `data/buildings.yaml`

The issues with list that should be "tuples" is addressed. YAML doesn't
really have tuples. Therefore, I have added the building IDs separated
by ", " so that they are easily manipulated.

I have modified the unit test
`test_all_required_buildings_are_building_ids()` to account for this so
that errors are not introduced. By convetion, halls are always
introduced first.

Lastly, #90 mentions that the required buildings for Stables are wrong
because each line is missing the "village_hall" building. But this is
not correct. Each line needs to list the immediate dependencies, not the
recursive ones. The Stables depend on Farm, Large farm, Vineyard and
Fishing village just like the file already states.

### Changes to the `Building` class

The init and post init for the field are modified accordingly. The post
init takes care of transforming the list of strings containing building
IDs separated by ", " into a `list[tuple[str, ...]]`.

Two new staticmethods are introduced:

- `_format_building()` for formatting `Building(id = "...")`
- `_format_none()` for formatting None values

These two methods are applied where applicable (like the required geo
and rss, or the replaces). `_format_building()` is not applied to the
title because that one is an exception with a different format.

The `_building_required_building()` method is updated to show the
buildings, correctly printing AND and OR conditions where applicable.
Examples of this are added to `./building_example.py` for reference.

The height of the display is adjusted to respond correctly to different
buildings with different numbers of rows of dependencies.

### Other related updates

The unit tests are expanded with some building scenarios. Maybe it would
be good to, eventually, add all building scenarios (since there's only a
limited number of them) to ensure there are no faulty modifications to
the YAML file. This tests are classified under `mark.building_scenarios`
so the new mark is added to the `pytest.ini` file.

The `./README.md` file is updated to show so that the image of the
example shows the building with the dependency.

## Additional bug

Found a bug in the method `_building_required_rss()` which was causing
the building to be displayed with errors like `Required resource:
GeoFeature.ORE`. Fixed this here too.

## Other updates

Modified the labels:

- Required GeoFeature -> Required geo. feature
- Required Resource -> Required resource