Modeling of Fort cities (#102)
This PR addresses the modeling of the Fort cities in the Germania
campaign.

While re-playing the campaign I confirmed that these cities do count
towards victory. They also contribute 150 to the storage capacity of
each rss (each Fort). Therefore, it is important to model them for the
`Kingdom` class to work properly for this campaign.

## Changes

This PR introduces the `is_fort` boolean to the `./data/cities.yaml` and
the `City` class. It also introduces the "fort" building. Its name is
set to "Small Fort". This can be confusing with the already existing
Small Fort, but it is what the game calls it.

The `City` class is updated so that, if a city is marked as `is_fort` in
the YAML file, it will load this information and validate from there (0
additional buildings).

The unit tests for the "cities data" are adjusted to include this new
key in the file. New city scenario tests are added for the forts.