Add custom exceptions (#104)
This PR adds custom exceptions.

A hierarchy of exceptions is created. The `LegionError` is the base
exception for all custom exceptions in the project. The next level is a
class for all exceptions in each module. Then come the actual exceptions
that will be used in the different situations.

This is the hierarchy as of right now:

```txt
LegionError
  |_ BuildingError
  |    |_ UnknownBuildingError
  |    |_ TooManyWorkersError
  |    |_ InsufficientNumberOfWorkersError
  |    |_ NegativeNumberOfWorkersError
  |_ CityError
  |    |_ NoCityHallError
  |    |_ TooManyHallsError
  |    |_ FortsCannotHaveBuildingsError
  |    |_ TooManyBuildingsError
  |    |_ NoGarrisonFoundError
  |_ DisplayError
  |_ EffectsError
  |_ GeoFeaturesError
  |_ KingdomError
  |    |_ DuplicatedCityError
  |    |_ CitiesFromMultipleCampaignsError
  |_ ResourcesError
  |_ ScenarioError
```

---
This PR also solves a bug that I found while working on this. Currently,
users can pass negative values to the methods for assigning workers in
the `Building` class. I have added a validation to ensure `qty >= 0`. If
`qty < 0`, a `NegativeNumberOfWorkersError` is raised.