Update `Scenario` examples (#107)
This PR updates the `scenario_example.py` script to depict difference
cities where the question that the `Scenario` class is trying to answer
is more pressing.

The first two examples are cities where more than 6 production buildings
can be built and staffed, but this would come at a cost of the "Basilica
+ Guild" combo. Which configurations produce more? These scenarios
highlight the trade-off between more base production Vs more
productivity bonus.

The third is an atempt to pill up as many production bonuses as possible
and showcase what happens with the rss production.