Add `_CityValidator` class (#109)
This PR introduces the `_CityValidator` class. The validations that
already exist are moved to this class (`_validate_halls()` and
`_validate_number_of_buildings()`).

This class will be the orchestrator of the city validation process.
Eventually, it will take care of all validations. This class is needed
for #37, but other building blocks (like #15) are still missing to
complete it.