Adds the `City.hall` property (#114)
This PR adds the `City.hall` property. This is a special building. All
cities must have a hall. Also, the "level" of the hall determines how
many other buildings the city can have, and different buildings are
enabled or blocked (yes, blocked, see #113 for more on this) by the
different hall "levels".

Other operations, like validating the city (#37) will become easier by
using the hall information. Therefore, it makes sense to avoid having to
repeatedly call `City.get_hall()` constantly.

Since this PR adds the hall as a property of the city, it now marks the
`get_hall()` method as internal (`_get_hall()`) and removes all uses of
it. The hall property gets filled during `__post_init__` and from there
on only the `City.hall` property is used.

The property is not just the building ID, but the entire `Building`
instance. This allows for different methods/functions and logics to
consume from it what they need (like storage values for example).