Removes city validator class (#115)
This PR removes the `_CityValidator` class. Though the idea of having a
dedicated class that takes care of validation is good, this
implementation is problematic.

The first problem, a conceptual one, is that the `_CityValidator` class
is not taking care of all the validations. The `campaign` and `name`
property also need to be validated and those are being taken care of by
the `City` class indirectly.

Second, having the validation occur in `__post_init__` means that the
potentially-invalid city has already been init'ed by the time the
validation occurs.

This PR basically reverts #109 and #114. The `City.hall` property is
precisely one of the places where the issues created by the order of
operations arise. The field is being referenced in the other methods of
the `City` class but it does not yet exist, because it is generated
during `City.__post_init__()`. This cannot really be changed because the
city configuration first needs to be validated, before we can say that
there's a hall for that city.

---
Potentially, one possible way of implementing this would be with a
hierarchy of classes. If `_CityValidator -> City` then when creating an
instance of the `City` class, the `super().__init__()` method can be
invoked. The `_CityValidator.__init__()` method will then take care of
validating. If valid, all properties of the `City` will be assigned. But
this implementation would need to be tested.