Refactor the `City` class (#117)
This PR refactors the `City` class. It deprecates the use of a dataclass
in favor of a regular Python class.

The rationale behind this decision is that the class' logics, especially
its init logic, has surpassed what the dataclass can do efficiently.
This has been clearly shown by #115 and #116. Regarding #115, since this
refactoring solves the issues, the `City.hall` field is re-introduced.