Add building counts validation (#118)
This PR adds the building counts validation.

This PR introduces the method
`City._calculate_allowed_building_counts()` which contains the logics
for determining the maximum number of instances of buildings that a city
can have. This is calculated for each building and it takes into
consideration the city's resources, its geo features, etc.

These max. number of building values are then used in the
`City._validate_number_of_buildings()` method to not only flag cities
with too many different types of buildings, but also raise errors when
the number of building instances exceeds the number of allowed building
instances for that city.

A whole slew of different unit tests are added to test that the allowed
counts are correctly computed for different cases. This also includes
testing that some "impossible scenarios" do raise errors.

---
This PR concludes the city validation (#37), and it achieves it without
the complication of bilding a graph (#15).