Correct counts for hunters' lodges (#120)
This PR corrects the allowed building counts for the HLs as described in
https://github.com/daczarne/legion/issues/113#issuecomment-3253865075.
In short, no new HLs can be built once the city reaches "city_hall", but
previously existing ones don't get automatically destroyed. So that
configuration of "city_hall" + up to 6 HLs is indeed possible (if the
other conditions are met).