Updates class vars and dunder methods for the `City` class (#123)
This PR updates the class variables and the dunder methods for the
`City` class.

For the dunder methods, the only change is that they are now sorted
alphabetically.

For the class vars, the var `POSSIBLE_CITY_HALLS` is renamed to
`POSSIBLE_HALLS`. This is done because "city_hall" is one of the
possible halls. This name avoids potential confusion.