Add workers' distribution (#130)
This PR introduces the mechanism for staffing buildings. Since this is
an active decision that the player needs to make while playing, a new
argument is added to the `City` class called `staffing_strategy`. For
now, this argument can only be a string and it is limited to four
possible values:

- `production_first`: will staff the production buildings first
- `production_only`: will only staff the production buildings.
Effects-buildings will not receive workers, even if there are more
workers available.
- `effects_first`: will staff the effects-buildings first
- `effects_only`: will only staff the effects buildings. Production
buildings will not receive workers, even if there are more workers
available.

Buildings are classified between "production" and "effects". Production
buildings are those for which adding workers increases their resource
output (farms, mines, lumber mills, etc). Effects-buildings are those
for which adding workers increases the effect bonuses (troop training,
population growth, intelligence). Since there are only two categories,
the program is only concerned with classifying the production buildings.
All other buildings are grouped in the non-production buildings.