Adds methods to the `Kingdom` class (#139)
This PR adds the `Kingdom.get_city()` method. This method take the name
of a city in the kingdom and returns the city object. If the city is not
found it raises an error. This is unorthodox, but in this case, because
the kingdom object has been created from a valid list of cities, the
only way a city can be missing is if the city does not belong to the
kingdom.

To validate that a given city indeed belongs to the kingdom, the
`has_city()` method is also added.