Adds validation of available building spots (#140)
This PR fixes a subtle bug in the city validation process. In short, the
city as a whole needs to be validated, not just individual buildings.
See #138 for more details.

This PR solves the problem by adding the
`_validate_empty_building_spots()` method. This method counts how many
of the buildings supplied to the `City` constructor will require an
empty or "emptyable" building-spot, and how many empty or "emptyable"
building-spots are there in the city. If more spots are required than
are available, an error is raised.

This method accounts for building-spots blocked by geo features, halls,
and supply dumps. Some of these are of course redundant as there are
other validations that already specialise on them. But including them
here makes the method logically consistent and protects the validation
process from the hypothetical future-removal of other validation steps.

This PR adds all necessary updates to the documentation, examples, and
unit tests for/related to this new method.