Adjusts ruff configuration
Disables rule PT019 because this rule produces false negatives. It is
demanding that `@pytest.mark.usefixtures` be used in situationes in
which using it would break the tests.

Adds `[tool.ruff.lint.per-file-ignores]` for `tests/*` files. These are
rules that I want to have for scripts and modules, but are incompatible
or an overkill in tests environments.

"D1" rules. These rules about docstrings just don't make sense to have
in with tests. Documentation is a maintenance cost too.

"PLR2004" rule. This forbids the use of "magic" numbers. But unit tests
that check that the results of calculations are correct are full of
these numbers. Each value is clearly the "expected value of a
calculation" and it will only be used once. Therefore, it doesn't make
sense to declare varaibles for each one of them.

"F841". This rule checks for unused variables. While in scripts and
modules this is definetly a good check to have, in tests that check for
failures with context managers having this variables declared is a
personal preference. Since the calls fail, the values never get
assigned, and therefore, the variable is just for redability purposes.

"SLF001" this rule checks access of private methods outside of the
class. Good to have in modules and scripts. But just because a method is
private does not mean that it does not need to be tested! Therefore, in
test contexts private methods will always be used.